flatblog/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flatblog"
version = "0.1.0"
description = "Standalone flat-file AI blogging system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9",
    "rich>=13",
    "jinja2>=3.1",
    "markdown-it-py>=3",
    "pyyaml>=6",
    "paramiko>=3",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[project.scripts]
flatblog = "flatblog.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/flatblog"]
